Drop missing values at columns that match a given pattern

Provides a simple yet efficient way to drop missing values("NA"s) at columns that match a given pattern.

drop_na_at(
  df,
  pattern_type = "contains",
  pattern = NULL,
  case_sensitive = FALSE,
  ...
)

Arguments

df
A data.frame object

pattern_type
One of "contains", "ends_with" or "starts_with"

pattern
The type of pattern to use when matching the pattern_type. The pattern is case sensitive

case_sensitive
Defaults to FALSE. Patterns are case insensitive if TRUE

...
Other params to other methods

Value

A data.frame object containing only columns that match the given pattern with the missing values removed.

Examples

head(drop_na_at(airquality,pattern_type = "starts_with","O"))

#>   Ozone
#> 1    41
#> 2    36
#> 3    12
#> 4    18
#> 5    28
#> 6    23