Sort Variables according to missingness

Provides a useful way to sort the variables(columns) according to their missingness.

sort_by_missingness(x, sort_by = "counts", descending = FALSE, ...)

Arguments

x
A valid R `object` for which `na_counts` are needed.

sort_by
One of counts or percents. This determines whether the results are sorted by counts or percentages.

descending
Logical. Should missing values be sorted in decreasing order ie largest to smallest? Defaults to FALSE.

...
Other arguments to specific functions. See "See also below"

Value

A `data.frame` object sorted by number/percentage of missing values

See also

Examples

sort_by_missingness(airquality, sort_by = "counts")

#>   variable percent
#> 1     Wind       0
#> 2     Temp       0
#> 3    Month       0
#> 4      Day       0
#> 5  Solar.R       7
#> 6    Ozone      37

# sort by percents
sort_by_missingness(airquality, sort_by="percents")

#>   variable   percent
#> 1     Wind  0.000000
#> 2     Temp  0.000000
#> 3    Month  0.000000
#> 4      Day  0.000000
#> 5  Solar.R  4.575163
#> 6    Ozone 24.183007

# descending order
sort_by_missingness(airquality, descend = TRUE)

#>   variable percent
#> 1    Ozone      37
#> 2  Solar.R       7
#> 3     Wind       0
#> 4     Temp       0
#> 5    Month       0
#> 6      Day       0