Recode Values as NA if they meet defined criteria

Recode Values as NA if they meet defined criteria

recode_as_na_for(df, criteria = "gt", value = 0, subset_cols = NULL)

Arguments

df

A data.frame object to manipulate

criteria

One of gt,gteq,lt,lteq to define greater than, greater than or equal to, less than or less than or equal to.

value

The value to convert to `NA`. We can for instance change "n/a" to `NA` or any other value.

subset_cols

An optional character vector for columns to manipulate.

Value

A data.frame object with the required changes.

Examples

recode_as_na_for(airquality,value=36, criteria = "gteq",
subset_cols = c("Ozone","Solar.R"))
#>     Ozone Solar.R Wind Temp Month Day
#> 1      NA      NA  7.4   67     5   1
#> 2      NA      NA  8.0   72     5   2
#> 3      12      NA 12.6   74     5   3
#> 4      18      NA 11.5   62     5   4
#> 5      NA      NA 14.3   56     5   5
#> 6      28      NA 14.9   66     5   6
#> 7      23      NA  8.6   65     5   7
#> 8      19      NA 13.8   59     5   8
#> 9       8      19 20.1   61     5   9
#> 10     NA      NA  8.6   69     5  10
#> 11      7      NA  6.9   74     5  11
#> 12     16      NA  9.7   69     5  12
#> 13     11      NA  9.2   66     5  13
#> 14     14      NA 10.9   68     5  14
#> 15     18      NA 13.2   58     5  15
#> 16     14      NA 11.5   64     5  16
#> 17     34      NA 12.0   66     5  17
#> 18      6      NA 18.4   57     5  18
#> 19     30      NA 11.5   68     5  19
#> 20     11      NA  9.7   62     5  20
#> 21      1       8  9.7   59     5  21
#> 22     11      NA 16.6   73     5  22
#> 23      4      25  9.7   61     5  23
#> 24     32      NA 12.0   61     5  24
#> 25     NA      NA 16.6   57     5  25
#> 26     NA      NA 14.9   58     5  26
#> 27     NA      NA  8.0   57     5  27
#> 28     23      13 12.0   67     5  28
#> 29     NA      NA 14.9   81     5  29
#> 30     NA      NA  5.7   79     5  30
#> 31     NA      NA  7.4   76     5  31
#> 32     NA      NA  8.6   78     6   1
#> 33     NA      NA  9.7   74     6   2
#> 34     NA      NA 16.1   67     6   3
#> 35     NA      NA  9.2   84     6   4
#> 36     NA      NA  8.6   85     6   5
#> 37     NA      NA 14.3   79     6   6
#> 38     29      NA  9.7   82     6   7
#> 39     NA      NA  6.9   87     6   8
#> 40     NA      NA 13.8   90     6   9
#> 41     NA      NA 11.5   87     6  10
#> 42     NA      NA 10.9   93     6  11
#> 43     NA      NA  9.2   92     6  12
#> 44     23      NA  8.0   82     6  13
#> 45     NA      NA 13.8   80     6  14
#> 46     NA      NA 11.5   79     6  15
#> 47     21      NA 14.9   77     6  16
#> 48     NA      NA 20.7   72     6  17
#> 49     20      NA  9.2   65     6  18
#> 50     12      NA 11.5   73     6  19
#> 51     13      NA 10.3   76     6  20
#> 52     NA      NA  6.3   77     6  21
#> 53     NA      NA  1.7   76     6  22
#> 54     NA      NA  4.6   76     6  23
#> 55     NA      NA  6.3   76     6  24
#> 56     NA      NA  8.0   75     6  25
#> 57     NA      NA  8.0   78     6  26
#> 58     NA      NA 10.3   73     6  27
#> 59     NA      NA 11.5   80     6  28
#> 60     NA      31 14.9   77     6  29
#> 61     NA      NA  8.0   83     6  30
#> 62     NA      NA  4.1   84     7   1
#> 63     NA      NA  9.2   85     7   2
#> 64     32      NA  9.2   81     7   3
#> 65     NA      NA 10.9   84     7   4
#> 66     NA      NA  4.6   83     7   5
#> 67     NA      NA 10.9   83     7   6
#> 68     NA      NA  5.1   88     7   7
#> 69     NA      NA  6.3   92     7   8
#> 70     NA      NA  5.7   92     7   9
#> 71     NA      NA  7.4   89     7  10
#> 72     NA      NA  8.6   82     7  11
#> 73     10      NA 14.3   73     7  12
#> 74     27      NA 14.9   81     7  13
#> 75     NA      NA 14.9   91     7  14
#> 76      7      NA 14.3   80     7  15
#> 77     NA      NA  6.9   81     7  16
#> 78     35      NA 10.3   82     7  17
#> 79     NA      NA  6.3   84     7  18
#> 80     NA      NA  5.1   87     7  19
#> 81     NA      NA 11.5   85     7  20
#> 82     16       7  6.9   74     7  21
#> 83     NA      NA  9.7   81     7  22
#> 84     NA      NA 11.5   82     7  23
#> 85     NA      NA  8.6   86     7  24
#> 86     NA      NA  8.0   85     7  25
#> 87     20      NA  8.6   82     7  26
#> 88     NA      NA 12.0   86     7  27
#> 89     NA      NA  7.4   88     7  28
#> 90     NA      NA  7.4   86     7  29
#> 91     NA      NA  7.4   83     7  30
#> 92     NA      NA  9.2   81     7  31
#> 93     NA      NA  6.9   81     8   1
#> 94      9      24 13.8   81     8   2
#> 95     16      NA  7.4   82     8   3
#> 96     NA      NA  6.9   86     8   4
#> 97     35      NA  7.4   85     8   5
#> 98     NA      NA  4.6   87     8   6
#> 99     NA      NA  4.0   89     8   7
#> 100    NA      NA 10.3   90     8   8
#> 101    NA      NA  8.0   90     8   9
#> 102    NA      NA  8.6   92     8  10
#> 103    NA      NA 11.5   86     8  11
#> 104    NA      NA 11.5   86     8  12
#> 105    28      NA 11.5   82     8  13
#> 106    NA      NA  9.7   80     8  14
#> 107    NA      NA 11.5   79     8  15
#> 108    22      NA 10.3   77     8  16
#> 109    NA      NA  6.3   79     8  17
#> 110    23      NA  7.4   76     8  18
#> 111    31      NA 10.9   78     8  19
#> 112    NA      NA 10.3   78     8  20
#> 113    21      NA 15.5   77     8  21
#> 114     9      NA 14.3   72     8  22
#> 115    NA      NA 12.6   75     8  23
#> 116    NA      NA  9.7   79     8  24
#> 117    NA      NA  3.4   81     8  25
#> 118    NA      NA  8.0   86     8  26
#> 119    NA      NA  5.7   88     8  27
#> 120    NA      NA  9.7   97     8  28
#> 121    NA      NA  2.3   94     8  29
#> 122    NA      NA  6.3   96     8  30
#> 123    NA      NA  6.3   94     8  31
#> 124    NA      NA  6.9   91     9   1
#> 125    NA      NA  5.1   92     9   2
#> 126    NA      NA  2.8   93     9   3
#> 127    NA      NA  4.6   93     9   4
#> 128    NA      NA  7.4   87     9   5
#> 129    32      NA 15.5   84     9   6
#> 130    20      NA 10.9   80     9   7
#> 131    23      NA 10.3   78     9   8
#> 132    21      NA 10.9   75     9   9
#> 133    24      NA  9.7   73     9  10
#> 134    NA      NA 14.9   81     9  11
#> 135    21      NA 15.5   76     9  12
#> 136    28      NA  6.3   77     9  13
#> 137     9      24 10.9   71     9  14
#> 138    13      NA 11.5   71     9  15
#> 139    NA      NA  6.9   78     9  16
#> 140    18      NA 13.8   67     9  17
#> 141    13      27 10.3   76     9  18
#> 142    24      NA 10.3   68     9  19
#> 143    16      NA  8.0   82     9  20
#> 144    13      NA 12.6   64     9  21
#> 145    23      14  9.2   71     9  22
#> 146    NA      NA 10.3   81     9  23
#> 147     7      NA 10.3   69     9  24
#> 148    14      20 16.6   63     9  25
#> 149    30      NA  6.9   70     9  26
#> 150    NA      NA 13.2   77     9  27
#> 151    14      NA 14.3   75     9  28
#> 152    18      NA  8.0   76     9  29
#> 153    20      NA 11.5   68     9  30